Double precision, relative emptiness tolerance, degenerate-triangle guard
Positions, normals and the radius are double. The empty-ball test treated
a point as inside when its squared distance to the centre was below
radius^2 - 1e-4, an absolute margin that is 4% of the ball on a radius of
0.05 and larger than the ball itself below 0.01: on a lattice torus with
radius 0.05 it left 234 boundary edges, and on a scan in metres it accepted
anything. The margin is now a relative 1e-9 of the radius, and the
circumcentre computation refuses a degenerate triangle instead of dividing
by zero.

On the comparison inputs the torus is now closed (19200 triangles, no
boundary), the other cases are within a few triangles of before, and the
time is unchanged.

+++ src/driver/main.cpp
@@ -11,7 +11,7 @@ int main(int argc, char* argv[]) try {
 	}
 
 	const std::filesystem::path inputFile = argv[1];
-	const float radius = std::stof(argv[2]);
+	const double radius = std::stod(argv[2]);
 	const std::filesystem::path outputFile = argc == 4 ? argv[3] : inputFile.string() + ".off";
 
 	const auto points = bpa::loadPoints(inputFile);
